Add support for cics provisioners
