+++ galasa-managers-parent/galasa-managers-cicsts-parent/dev.galasa.cicsts.manager/src/main/java/dev/galasa/cicsts/internal/dse/DseCicsImpl.java
@@ -9,40 +9,16 @@ import dev.galasa.ProductVersion;
 import dev.galasa.cicsts.CicstsManagerException;
 import dev.galasa.cicsts.internal.CicstsManagerImpl;
 import dev.galasa.cicsts.internal.properties.DseVersion;
-import dev.galasa.cicsts.spi.ICicsRegionProvisioned;
+import dev.galasa.cicsts.spi.BaseCicsImpl;
 import dev.galasa.zos.IZosImage;
-import dev.galasa.zos.ZosManagerException;
 
-public class DseCicsImpl implements ICicsRegionProvisioned {
-
-    private final String cicsTag;
-    private final String applid;
-    private final IZosImage zosImage;
+public class DseCicsImpl extends BaseCicsImpl {
 
     private ProductVersion version;
 
-    private int lastTerminalId;
-
-    public DseCicsImpl(CicstsManagerImpl cicstsManager, String cicsTag, String imageTag, String applid)
+    public DseCicsImpl(CicstsManagerImpl cicstsManager, String cicsTag, IZosImage image, String applid)
             throws CicstsManagerException {
-        this.cicsTag = cicsTag;
-        this.applid = applid;
-
-        try {
-            this.zosImage = cicstsManager.getZosManager().getImageForTag(imageTag);
-        } catch (ZosManagerException e) {
-            throw new CicstsManagerException("Unable to locate zOS Image tagged " + imageTag, e);
-        }
-    }
-
-    @Override
-    public String getTag() {
-        return this.cicsTag;
-    }
-
-    @Override
-    public String getApplid() {
-        return this.applid;
+        super(cicstsManager, cicsTag, image, applid);
     }
 
     @Override
@@ -51,35 +27,21 @@ public class DseCicsImpl implements ICicsRegionProvisioned {
             return this.version;
         }
 
-        String versionString = DseVersion.get(this.cicsTag);
+        String versionString = DseVersion.get(this.getTag());
 
         if (versionString == null) {
-            throw new CicstsManagerException("The version was missing for DSE tag " + this.cicsTag);
+            throw new CicstsManagerException("The version was missing for DSE tag " + this.getTag());
         }
 
         try {
             this.version = ProductVersion.parse(versionString);
         } catch (Exception e) {
-            throw new CicstsManagerException("Invalid version string for DSE tag " + this.cicsTag + ", format should be 0.0.0", e);
+            throw new CicstsManagerException("Invalid version string for DSE tag " + this.getTag() + ", format should be 0.0.0", e);
         }
 
         return this.version;
     }
 
-    @Override
-    public IZosImage getZosImage() {
-        return this.zosImage;
-    }
 
-    @Override
-    public String toString() {
-        return "CICS Region[" + this.applid + "]";
-    }
-
-    @Override
-    public String getNextTerminalId() {
-        lastTerminalId++;
-        return this.applid + "_" + Integer.toString(lastTerminalId);
-    }
 
 }
