+++ galasa-managers-parent/galasa-managers-cicsts-parent/dev.galasa.cicsts.manager/src/main/java/dev/galasa/cicsts/spi/ICicsRegionProvisioner.java
@@ -11,9 +11,12 @@ import java.util.List;
 import javax.validation.constraints.NotNull;
 
 import dev.galasa.ManagerException;
+import dev.galasa.framework.spi.ResourceUnavailableException;
 
 public interface ICicsRegionProvisioner {
 
     ICicsRegionProvisioned provision(@NotNull String cicsTag, @NotNull String imageTag, @NotNull List<Annotation> annotations) throws ManagerException;
+
+    void cicsProvisionGenerate() throws ManagerException, ResourceUnavailableException;
 
 }
