+++ galasa-managers-parent/galasa-managers-cicsts-parent/dev.galasa.cicsts.manager/src/main/java/dev/galasa/cicsts/spi/ICicstsManagerSpi.java
@@ -9,11 +9,19 @@ import java.util.List;
 
 import javax.validation.constraints.NotNull;
 
+import dev.galasa.ProductVersion;
+
 public interface ICicstsManagerSpi {
 
     void registerProvisioner(ICicsRegionProvisioner provisioner);
 
     @NotNull
     List<ICicsRegionLogonProvider> getLogonProviders();
+
+    @NotNull
+    String getProvisionType();
+
+    @NotNull
+    ProductVersion getDefaultVersion();
 
 }
